extract send functions
